old/notes4_reposetup/requirements.txt
cleandoc>=0.0.5
PyYAML
types-PyYAML
GitPython
pytest
pipreqs

